[deployment] reworked redditbot to be split up
