+++ redditbot.py
@@ -1,45 +1,29 @@
+"""
+Reddit bot that checks mentions, and launches brain.py when new mentions are detected
+"""
+
 import argparse
-import datetime
 import os
 from pprint import pprint
 import re
-import string
-import sys
 import time
-from collections import OrderedDict
 
-import pandas as pd
 import praw
 import prawcore
-import spacy
 from praw.models import Comment, Submission
-from tinydb import Query, TinyDB
-import validators
+from tinydb import TinyDB
 import yaml
 
-from argmatcher import ArgMatcher
-from response_templates import END_TEMPLATE, FAILURE_COMMENT, FAILURE_PM
+import subprocess
 
 
 def parse_args():
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "--run-once",
-        help="Run only once, checking mentions",
-        action="store_true",
-        default=False,
-    )
-    parser.add_argument(
         "--check-replied",
         help="Check if we have replied already first, only necessary once",
         action="store_true",
         default=False,
-    )
-    parser.add_argument(
-        "--limit",
-        help="Maximum number of mentions to check (default=-1, unlimited)",
-        type=int,
-        default=-1,
     )
     parser.add_argument(
         "--config",
@@ -55,9 +39,6 @@ def parse_args():
     )
     args = parser.parse_args()
 
-    if args.limit <= 0:
-        args.limit = None
-
     assert os.path.isfile(args.config)
     assert os.path.isfile(args.log_db)
 
@@ -70,36 +51,17 @@ def load_config_yaml(file):
     return config
 
 
-def load_myth_links(file):
-    df = pd.read_csv(file)
-    df = df.fillna("")
-    return OrderedDict(
-        {k: v for k, v in zip(df["Title"].values, df["Link"].values) if v}
-    )
-
-
 class MentionsBot:
     def __init__(
         self,
-        argmatch,
         config,
-        db,
+        config_file,
+        db_file,
     ):
         self.config = config
+        self.config_file = config_file
         self.reddit = praw.Reddit(check_for_async=False, **config["user_info"])
         self.inbox = praw.models.Inbox(self.reddit, _data={})
-        self.argmatch = argmatch
-
-        ## Config/Thresholds for standard matching
-        self.n_neighbors = int(config["n_neighbors"])
-        self.threshold = float(config["threshold"])
-        self.certain_threshold = float(config["certain_threshold"])
-
-        ## Config/Thresholds for matching with a hint
-        self.hint_n_neighbors = int(config["hint_n_neighbors"])
-        self.hint_arg_threshold = float(config["hint_arg_threshold"])
-        self.hint_threshold = float(config["hint_threshold"])
-        self.hint_certain_threshold = float(config["hint_certain_threshold"])
 
         self.whitelisted_subreddits = set(config["whitelisted"])
         self.blacklisted_subreddits = set(["suicidewatch", "depression"]).union(
@@ -113,25 +75,19 @@ class MentionsBot:
             [s.lower() for s in self.blacklisted_subreddits]
         )
 
-        self.db = db
-        self.replied = self.fill_replied(self.db)
+        self.db = TinyDB(db_file)
+        self.db_file = db_file
+        self.replied = set()
+        self.fill_replied()
 
-        self.alphabet = string.ascii_letters
-
-        self.END_TEMPLATE = END_TEMPLATE
-        self.FAILURE_COMMENT = FAILURE_COMMENT
-        self.FAILURE_PM = FAILURE_PM
-
-    def fill_replied(self, db):
+    def fill_replied(self):
         """
-        Returns a list of all parent and mention ids found in the log DB
+        Populates the replied set
         """
-        replied = set()
-        for entry in db.all():
-            replied.add(entry["mention_id"])
+        for entry in self.db.all():
+            self.replied.add(entry["mention_id"])
             if "parent_id" in entry:
-                replied.add(entry["parent_id"])
-        return replied
+                self.replied.add(entry["parent_id"])
 
     def clear_already_replied(self):
         """
@@ -171,57 +127,9 @@ class MentionsBot:
                             self.replied.add(parent.id)
                             self.db.insert(reply_info)
 
-    def format_response(self, resps):
+    def check_mentions(self, limit=None):
         """
-        Formatting responses given from the argument matcher
-        """
-        args = OrderedDict({})
-        for r in resps:
-            inp = r["input_sentence"]
-            arg = r["matched_argument"]
-            passage = r["reply_text"]
-            sim = r["similarity"]
-            link = r["link"]
-
-            if arg not in args:
-                args[arg] = {
-                    "passage": passage,
-                    "quotes": [inp],
-                    "sim": sim,
-                    "link": link,
-                }
-            else:
-                args[arg]["quotes"].append(inp)
-                if args[arg]["sim"] < sim:
-                    # replace the passage if this sentence is better matched
-                    args[arg]["sim"] = sim
-                    args[arg]["passage"] = passage
-
-        replies = []
-        for i, arg in enumerate(args):
-            parts = []
-            quotes = "".join(
-                [">{} \n\n".format(q) for q in args[arg]["quotes"]]
-            ) + "> ^(({})^) \n\n".format(self.alphabet[i])
-            passage = args[arg]["passage"]
-            if i < len(args) - 1:
-                # Only add dividers between args
-                passage += "\n\n --- \n\n"
-            parts.append(quotes)
-            parts.append(passage)
-            arglist = "({}): {}".format(self.alphabet[i], arg)
-            link = args[arg]["link"]
-            if validators.url(link):
-                arglist = "[({}): {}]({})".format(self.alphabet[i], arg, link)
-
-            parts.append(self.END_TEMPLATE.format(arglist))
-            replies.append("\n".join(parts))
-        return replies
-
-    def reply_mentions(self, limit=None):
-        """
-        Main functionality. Go through mentions and reply to parent comments
-        Uses persentence argmatcher
+        Check for mentions, run brain.py if a valid mention is detected
         """
         for mention in self.inbox.mentions(limit=limit):
 
@@ -257,135 +165,16 @@ class MentionsBot:
 
             # Proceed if mention has not been dealt with
             if mention.id not in self.replied:
-                if isinstance(mention, Comment):
-                    parent = mention.parent()
-                    reply_info["parent_id"] = parent.id
-                    reply_info["parent_username"] = (
-                        parent.author.name if parent.author else None
-                    )
+                self.launch_brain()
+                self.replied.add(mention.id)
 
-                    # Check if parent has been handled (in case of multiple mentions)
-                    if parent.id in self.replied:
-                        reply_info["outcome"] = "Parent already replied to"
-                        self.replied.add(mention.id)
-                        self.db.insert(reply_info)
-                        continue
-
-                    try:
-                        if isinstance(parent, Comment):
-                            input_text = self.remove_usernames(parent.body)
-                        elif isinstance(parent, Submission):
-                            input_text = self.remove_usernames(
-                                ".".join([parent.title, parent.selftext])
-                            )
-                        else:
-                            input_text = None
-                    except:
-                        input_text = None
-
-                    reply_info["input_text"] = input_text
-
-                    if input_text:
-                        input_text = self.replace_newlines(input_text)
-                        mention_hints = self.remove_usernames(mention.body).replace(
-                            ",", "."
-                        )
-                        resps = self.argmatch.match_text(
-                            input_text,
-                            threshold=self.threshold,
-                            certain_threshold=self.certain_threshold,
-                            N_neighbors=self.n_neighbors,
-                        )
-
-                        if mention_hints:
-                            # Use mention hints to match arguments
-                            reply_info["mention_hints"] = mention_hints
-
-                            # This step looks at the mention hint, and gets the arglabels hinted
-                            # Hint arg threshold is low since we expect hint to be obvious
-                            # TODO: look into matching only with argument titles
-                            hint_resps = self.argmatch.match_text(
-                                mention_hints,
-                                threshold=self.hint_arg_threshold,
-                                certain_threshold=0.9,  # Irrelevant as N_neighbors=1
-                                N_neighbors=1,
-                                return_reply=False,
-                            )
-
-                            reply_info["hint_responses"] = hint_resps
-
-                            arg_labels = set(
-                                [r["matched_arglabel"] for r in hint_resps]
-                            )
-                            r_arg_labels = set([r["matched_arglabel"] for r in resps])
-
-                            # Check only the hinted args which aren't matched already
-                            arg_labels = arg_labels - r_arg_labels
-
-                            if arg_labels:
-                                # Pass arg_labels to match_text, restricting to hinted args
-                                hinted_resps = self.argmatch.match_text(
-                                    input_text,
-                                    arg_labels=arg_labels,
-                                    threshold=self.hint_threshold,
-                                    certain_threshold=self.hint_certain_threshold,
-                                    N_neighbors=self.hint_n_neighbors,
-                                )
-
-                                # Adds remaining responses to hinted ones
-                                # Skips if matched to a hinted arg
-                                oldresps = resps
-                                resps = hinted_resps
-                                hinted_sents = [
-                                    r["input_sentence"] for r in hinted_resps
-                                ]
-                                for r in oldresps:
-                                    if r["input_sentence"] in hinted_sents:
-                                        # This means the sentence was matched up to a hint
-                                        continue
-                                    else:
-                                        resps.append(r)
-
-                    else:
-                        resps = []
-
-                    reply_info["responses"] = resps
-
-                    if resps:  # Found arg match(es)
-                        formatted_responses = self.format_response(resps)
-                        reply_info["full_reply"] = formatted_responses
-
-                        for response in formatted_responses:
-                            try:
-                                reply = parent.reply(response)
-                                print(response)
-                                reply_info[
-                                    "outcome"
-                                ] = "Replied with matched argument(s)"
-                                reply_info["reply_id"] = reply.id
-                            except prawcore.exceptions.Forbidden:
-                                reply_info[
-                                    "outcome"
-                                ] = "Found arguments but failed to reply: Forbidden"
-
-                    else:  # Failed to find arg match
-                        mention.reply(self.FAILURE_COMMENT)
-                        try:
-                            mention.author.message(
-                                "We couldn't find a response!",
-                                self.FAILURE_PM.format(
-                                    self.argmatch.prefilter(parent.body)
-                                ),
-                            )
-                        except:
-                            # PM-ing people sometimes fails, but this is not critical
-                            pass
-                        reply_info["outcome"] = "Failed to find any matched arguments"
-
-                    # Add both the mention and the parent to the replied list
-                    self.replied.add(mention.id)
-                    self.replied.add(parent.id)
-                    self.db.insert(reply_info)
+    def launch_brain(self):
+        """
+        Closes db, runs brain.py, opens db again
+        """
+        self.db.close()
+        p = subprocess.call(["python3", "brain.py", "--log-db", self.db_file, "--config", self.config_file])
+        self.db = TinyDB(self.db_file)
 
     def run(self, refresh_rate=600, timeout_retry=600, check_replied=True, limit=None):
         """
@@ -400,7 +189,7 @@ class MentionsBot:
 
         while True:
             try:
-                self.reply_mentions(limit=limit)
+                self.check_mentions(limit=limit)
                 print(
                     "{}\tReplied to mentions, sleeping for {} seconds...".format(
                         time.ctime(), refresh_rate
@@ -415,33 +204,6 @@ class MentionsBot:
                 )
                 time.sleep(timeout_retry)
 
-    def run_once(self, check_replied=False, limit=None):
-        """
-        Run the bot once, checking all new mentions
-        """
-        if check_replied:
-            print("Checking previous mentions to see if we have replied already...")
-            self.clear_already_replied()
-
-        self.reply_mentions(limit=limit)
-        print("Successfully checked and/or replied to mentions, exiting successfully")
-
-    @staticmethod
-    def remove_usernames(text):
-        """
-        Removes any /u/username or u/username strings
-        """
-        newtext = re.sub("\/u\/[A-Za-z0-9_-]+", "", text)
-        newtext = re.sub("u\/[A-Za-z0-9_-]+", "", newtext)
-        return newtext
-
-    @staticmethod
-    def replace_newlines(text):
-        """
-        Replaces newline symbols with periods
-        """
-        return text.replace("\n", ". ")
-
 
 if __name__ == "__main__":
     args = parse_args()
@@ -450,23 +212,13 @@ if __name__ == "__main__":
 
     refresh_rate = int(config["refresh_rate"])
 
-    nlp = spacy.load("en_core_web_lg")
-    nlp.add_pipe("universal_sentence_encoder", config={"model_name": "en_use_lg"})
-
-    db = TinyDB(args.log_db)
-
-    argm = ArgMatcher(nlp, None, None, preload=True)
     mb = MentionsBot(
-        argm,
         config,
-        db,
+        args.config,
+        args.log_db,
     )
 
-    if args.run_once:
-        mb.run_once(check_replied=args.check_replied, limit=args.limit)
-    else:
-        mb.run(
-            refresh_rate=refresh_rate,
-            check_replied=args.check_replied,
-            limit=args.limit,
-        )
+    mb.run(
+        refresh_rate=refresh_rate,
+        check_replied=args.check_replied,
+    )
